Refactor radius field type to Double in CampaignRequest and CampaignResponse; add CampaignService for CRUD operations

+++ backend/src/main/java/com/example/backend/dto/CampaignRequest.java
@@ -30,5 +30,5 @@ public class CampaignRequest {
 
     @NotNull(message = "Radius is mandatory")
     @Min(value = 1, message = "Radius must be at least 1km")
-    private Integer radius;
+    private Double radius;
 }
